refactor: move DeviceInfoProvider to app package
