+++ app/src/main/java/com/hilingual/di/ProviderModule.kt
@@ -1,6 +1,6 @@
 package com.hilingual.di
 
-import com.hilingual.core.common.provider.DeviceInfoProvider
+import com.hilingual.core.common.app.DeviceInfoProvider
 import com.hilingual.app.DeviceInfoProviderImpl
 import dagger.Binds
 import dagger.Module
