+++ core/common/src/main/java/com/hilingual/core/common/app/DeviceInfoProvider.kt
@@ -1,4 +1,4 @@
-package com.hilingual.core.common.provider
+package com.hilingual.core.common.app
 
 interface DeviceInfoProvider {
     fun getDeviceName(): String
