+++ presentation/mypage/src/main/java/com/hilingual/presentation/mypage/MyPageViewModel.kt
@@ -19,7 +19,7 @@ import android.net.Uri
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.hilingual.core.common.extension.onLogFailure
-import com.hilingual.core.common.provider.DeviceInfoProvider
+import com.hilingual.core.common.app.DeviceInfoProvider
 import com.hilingual.core.common.util.UiState
 import com.hilingual.data.auth.repository.AuthRepository
 import com.hilingual.data.user.repository.UserRepository
